pyproject.toml
[build-system]
requires = ["uv_build"]
build-backend = "uv_build"

[project]
name = "g4x_helpers"
version = "0.5.2"

description = "Python helpers for G4X."
urls = { repository = "https://github.com/Singular-Genomics/G4X-helpers" }
readme = "README.md"
license = { text = "MIT" }
authors = [
    { name = "Kenneth Gouin III", email = "kennethg@singulargenomics.com" },
    { name = "Florian Raths", email = "flo.raths@singulargenomics.com" },
]
requires-python = ">=3.10, <3.13"

classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
]

dependencies = [
    "geopandas ~= 1.0",
    "glymur ~= 0.12",
    "matplotlib ~= 3.7",
    "polars ~= 1.30",
    "protobuf ~= 4.25",
    "rasterio ~= 1.4",
    "scanpy ~= 1.10",
    "scikit-image ~= 0.25",
    "shapely ~= 2.1",
    "pyarrow ~= 16.1",
    "rich-click ~= 1.9",

]

[dependency-groups]
dev = [
    "ipykernel >= 6.30.1",
    "ipywidgets >= 8.1.7",
    "pre-commit ~= 4.2",
    "pytest ~= 8.4",
    "ruff == 0.13.1",    # pinned to match pre-commit version
]
docs = ["mkdocs ~= 1.6", "mkdocs-material ~= 9.6"]

[project.scripts]
g4x-helpers = "g4x_helpers.cli:cli"

[tool.uv]
required-version = ">=0.8.8"     # also update in Dockerfile
default-groups = ["dev", "docs"]

[tool.uv-ship]
release-branch = "main"
changelog-path = "CHANGELOG"
